Convervative garbage collector
A mark-sweep collector for the new C backend.  Conservative because
pointers and integers can't be distinguished in C, so we may hold on to
more heap the necessary if we treat something as a pointer that isn't.
Seems ok for our purposes (comparing against nicer implementations).
Mark the C backend as "experimental" for this reason.

One thing to think more about: should TAG_PTR be changed from 3 to 0?
Is it possible for a C pointer in the stack/registers to point to the
Elite heap?  If so, the GC should follow C pointers as well as Elite
pointers.